+++ great_expectations/datasource/fluent/pandas_azure_blob_storage_datasource.pyi
@@ -16,6 +16,7 @@ from typing import (
 from great_expectations._docs_decorators import public_api as public_api
 from great_expectations.compatibility import azure
 from great_expectations.compatibility.typing_extensions import override
+from great_expectations.core.sorters import SortersDefinition as SortersDefinition
 from great_expectations.core.util import AzureUrl as AzureUrl
 from great_expectations.datasource.fluent import _PandasFilePathDatasource
 from great_expectations.datasource.fluent.config_str import ConfigStr
@@ -33,9 +34,6 @@ from great_expectations.datasource.fluent.dynamic_pandas import (
     StorageOptions,
 )
 from great_expectations.datasource.fluent.interfaces import BatchMetadata
-from great_expectations.datasource.fluent.interfaces import (
-    SortersDefinition as SortersDefinition,
-)
 from great_expectations.datasource.fluent.interfaces import (
     TestConnectionError as TestConnectionError,
 )
