+++ great_expectations/datasource/fluent/pandas_dbfs_datasource.pyi
@@ -5,7 +5,9 @@ from typing import Hashable, Iterable, Literal, Optional, Sequence, Union
 
 from great_expectations._docs_decorators import public_api as public_api
 from great_expectations.compatibility.typing_extensions import override
-from great_expectations.datasource.fluent import PandasFilesystemDatasource, Sorter
+from great_expectations.core.sorters import Sorter
+from great_expectations.core.sorters import SortersDefinition as SortersDefinition
+from great_expectations.datasource.fluent import PandasFilesystemDatasource
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     DBFSDataConnector as DBFSDataConnector,
 )
@@ -17,9 +19,6 @@ from great_expectations.datasource.fluent.dynamic_pandas import (
     StorageOptions,
 )
 from great_expectations.datasource.fluent.interfaces import BatchMetadata
-from great_expectations.datasource.fluent.interfaces import (
-    SortersDefinition as SortersDefinition,
-)
 from great_expectations.datasource.fluent.interfaces import (
     TestConnectionError as TestConnectionError,
 )
