+++ great_expectations/datasource/fluent/pandas_filesystem_datasource.pyi
@@ -14,7 +14,9 @@ from typing import (
 
 from great_expectations._docs_decorators import public_api as public_api
 from great_expectations.compatibility.typing_extensions import override
-from great_expectations.datasource.fluent import Sorter, _PandasFilePathDatasource
+from great_expectations.core.sorters import Sorter
+from great_expectations.core.sorters import SortersDefinition as SortersDefinition
+from great_expectations.datasource.fluent import _PandasFilePathDatasource
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     FilesystemDataConnector as FilesystemDataConnector,
 )
@@ -26,9 +28,6 @@ from great_expectations.datasource.fluent.dynamic_pandas import (
     StorageOptions,
 )
 from great_expectations.datasource.fluent.interfaces import BatchMetadata
-from great_expectations.datasource.fluent.interfaces import (
-    SortersDefinition as SortersDefinition,
-)
 from great_expectations.datasource.fluent.interfaces import (
     TestConnectionError as TestConnectionError,
 )
