+++ great_expectations/datasource/fluent/spark_filesystem_datasource.pyi
@@ -6,12 +6,10 @@ from typing import ClassVar, Literal, Optional, Type, Union
 from great_expectations.compatibility.pyspark import (
     types as pyspark_types,
 )
+from great_expectations.core.sorters import SortersDefinition
 from great_expectations.datasource.fluent import BatchMetadata, _SparkFilePathDatasource
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     FilesystemDataConnector,
-)
-from great_expectations.datasource.fluent.interfaces import (
-    SortersDefinition,
 )
 from great_expectations.datasource.fluent.spark_file_path_datasource import (
     CSVAsset,
