+++ great_expectations/datasource/fluent/spark_google_cloud_storage_datasource.pyi
@@ -3,15 +3,13 @@ from logging import Logger
 from typing import Any, ClassVar, Literal, Optional, Type
 
 from great_expectations.compatibility import google
+from great_expectations.core.sorters import SortersDefinition
 from great_expectations.datasource.fluent import BatchMetadata, _SparkFilePathDatasource
 from great_expectations.datasource.fluent.config_str import (
     ConfigStr,
 )
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     GoogleCloudStorageDataConnector,
-)
-from great_expectations.datasource.fluent.interfaces import (
-    SortersDefinition,
 )
 from great_expectations.datasource.fluent.spark_file_path_datasource import (
     CSVAsset,
