+++ great_expectations/datasource/fluent/spark_s3_datasource.pyi
@@ -2,15 +2,13 @@ import re
 from logging import Logger
 from typing import Any, ClassVar, Literal, Optional, Type
 
+from great_expectations.core.sorters import SortersDefinition
 from great_expectations.datasource.fluent import BatchMetadata, _SparkFilePathDatasource
 from great_expectations.datasource.fluent.config_str import (
     ConfigStr,
 )
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     S3DataConnector,
-)
-from great_expectations.datasource.fluent.interfaces import (
-    SortersDefinition,
 )
 from great_expectations.datasource.fluent.spark_file_path_datasource import (
     CSVAsset,
