+++ great_expectations/datasource/fluent/sqlite_datasource.py
@@ -35,12 +35,13 @@ from great_expectations.datasource.fluent.sql_datasource import (
 if TYPE_CHECKING:
     # min version of typing_extension missing `Self`, so it can't be imported at runtime
 
+    from great_expectations.core.sorters import SortersDefinition
     from great_expectations.datasource.fluent.interfaces import (
         BatchMetadata,
         BatchRequestOptions,
         DataAsset,
-        SortersDefinition,
     )
+
 
 # This module serves as an example of how to extend _SQLAssets for specific backends. The steps are:
 # 1. Create a plain class with the extensions necessary for the specific backend.
