+++ tests/datasource/fluent/test_spark_filesystem_datasource.py
@@ -18,12 +18,12 @@ from great_expectations.core.partitioners import (
     PartitionerColumnValue,
     PartitionerYearAndMonth,
 )
+from great_expectations.core.sorters import SortersDefinition
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     FilesystemDataConnector,
 )
 from great_expectations.datasource.fluent.file_path_data_asset import _FilePathDataAsset
 from great_expectations.datasource.fluent.interfaces import (
-    SortersDefinition,
     TestConnectionError,
 )
 from great_expectations.datasource.fluent.spark_file_path_datasource import (
